+++ src/lib/sharedConstants.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024 F5, Inc.
+ * Copyright 2025 F5, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -73,6 +73,7 @@ module.exports = {
         RoutingBGP: '/tm/net/routing/bgp',
         RoutingPrefixList: '/tm/net/routing/prefix-list',
         SecurityAnalytics: '/tm/security/analytics/settings',
+        DeviceDOS: '/tm/security/dos/device-config',
         SelfIp: '/tm/net/self',
         SnmpAgent: '/tm/sys/snmp',
         SnmpCommunity: '/tm/sys/snmp/communities',