+++ test/integration/property/propertiesUser.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2024 F5, Inc.
+ * Copyright 2025 F5, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@ function assertUserClass(properties, options) {
 }
 
 describe('User', function testAuthentication() {
-    this.timeout(600000);
+    this.timeout(900000);
 
     it('userType regular', () => {
         const options = {